Fix package entry url with `?raw` flag
